+++ gcs/src/test/java/com/google/cloud/hadoop/fs/gcs/GoogleHadoopFileSystemConfigurationTest.java
@@ -129,7 +129,8 @@ public class GoogleHadoopFileSystemConfigurationTest {
           put("fs.gs.write.parallel.composite.upload.part.file.name.prefix", "");
           put("fs.gs.fadvise.request.track.count", 3);
           put("fs.gs.operation.move.enable", false);
-          put("fs.gs.write.rolling.checksum.enable", false);
+          put("fs.gs.write.rolling.checksum.enable", true);
+          put("fs.gs.storage.client.cache.enable", false);
         }
       };